+++ frontend/src/components/widgets/resource-table/ResourceTable.jsx
@@ -22,11 +22,8 @@ import { Table } from "@/components/ui/shims/antd-structure";
 import { Typography } from "@/components/ui/shims/antd-typography";
 
 import { formattedDateTime, timeAgo } from "../../../helpers/GetStaticData";
+import { resolveOwnerDisplay } from "../owner-display";
 import "./ResourceTable.css";
-
-// Service-account address minted by `create_api_user_for_key`. Its owner is a
-// platform API key, not a person, so the cell is labelled rather than named.
-const PLATFORM_KEY_EMAIL_DOMAIN = "@platform.internal";
 
 // Stable, distinct avatar swatch per owner (seeded on email/name) like the
 // design: a light pastel fill paired with a matching darker initial.
@@ -221,32 +218,11 @@ function ResourceTable({
   };
 
   const renderOwner = (item) => {
-    // owner_emails is earliest-first; [0] is the primary shown owner.
-    // Fall back to created_by_email so rows with no live OWNER membership
-    // (pre-backfill rows) don't render "Unknown".
-    const ownerEmails = item?.[ownerEmailsProp];
-    const rawEmail =
-      (Array.isArray(ownerEmails) ? ownerEmails[0] : undefined) ??
-      item?.created_by_email;
-    // Reached only when a platform key's creator has since been deleted, so no
-    // human can be named. Suppress the synthetic address rather than dress a
-    // machine identity up as a colleague.
-    const isPlatformKey = Boolean(
-      rawEmail?.endsWith(PLATFORM_KEY_EMAIL_DOMAIN),
+    const { email, name, extra } = resolveOwnerDisplay(
+      item,
+      sessionDetails,
+      ownerEmailsProp,
     );
-    const email = isPlatformKey ? undefined : rawEmail;
-    // "Me" must track the DISPLAYED owner, not the viewer's own membership —
-    // else a co-owner sees "Me" over the primary owner's avatar/email. Match on
-    // the shown email so the creator viewing their own resource still reads "Me".
-    const isMe = Boolean(email) && email === sessionDetails?.email;
-    let name = email?.split("@")[0] || "Unknown";
-    if (isPlatformKey) {
-      name = "Platform key";
-    } else if (isMe) {
-      name = "Me";
-    }
-    const extra =
-      item?.co_owners_count > 1 ? ` +${item.co_owners_count - 1}` : "";
     const initials = (email || name).slice(0, 2).toUpperCase();
     const swatch = colorForSeed(email || name);
